Break import cycle by moving createVoteableUnionType call to a hackier place
