+++ packages/lesswrong/server/vulcan-lib/apollo-server/initGraphQL.ts
@@ -3,7 +3,6 @@
 
 import { makeExecutableSchema } from 'apollo-server';
 import { getAdditionalSchemas, queries, mutations, getContext, getDirectives, getResolvers, getCollections } from '../../../lib/vulcan-lib/graphql';
-import { createVoteableUnionType } from '../../votingGraphQL';
 import {
   selectorInputTemplate,
   mainTypeTemplate,
@@ -409,8 +408,6 @@ const generateSchema = (collection: CollectionBase<DbObject>) => {
 
 
 export const initGraphQL = () => {
-  createVoteableUnionType();
-  
   const { schemaText, addedResolvers } = getTypeDefs();
   
   let allResolvers = deepmerge(
